requirements.txt
fastapi
uvicorn[standard]
playwright
pytesseract
pdf2image
PyMuPDF
Pillow
python-multipart
requests
